fixed crash on drag over new column

+++ src/components/TodoItem.js
@@ -10,9 +10,11 @@ function TodoItem(props) {
     const [{isDragging}, drag] = useDrag({
         item: { type: ItemTypes.TODO, desc: props.text, colSrc: props.colName, completed: props.completed },
         end: (item, monitor) => {
-            const result = monitor.getDropResult();
-            props.dispatch(AddTodo(item.desc, result.colName, item.completed));
-            props.dispatch(RemoveTodo(props.id, item.colSrc))
+            if (monitor.didDrop()) {
+                const result = monitor.getDropResult();
+                props.dispatch(AddTodo(item.desc, result.colName, item.completed));
+                props.dispatch(RemoveTodo(props.id, item.colSrc))
+            }
         },
         collect: monitor => ({
             isDragging: !!monitor.isDragging(),
